% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelFreePeptides.R
\name{readMaxQuantPeptideMeasurements}
\alias{readMaxQuantPeptideMeasurements}
\title{Load a MaxQuant 'peptides.txt' dataset.}
\usage{
readMaxQuantPeptideMeasurements(path)
}
\arguments{
\item{path}{Full path to the peptides.txt file.}
}
\value{
An object of class 'MaxQuantPeptidesResult'.
}
\description{
Load a MaxQuant 'peptides.txt' dataset.
}
