% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LabelFreeProteinGroups.R
\name{readMaxQuantProteinMeasurements}
\alias{readMaxQuantProteinMeasurements}
\title{Load a MaxQuant 'proteinGroups.txt' dataset.}
\usage{
readMaxQuantProteinMeasurements(path)
}
\arguments{
\item{path}{Full path to the proteinGroups.txt file.}
}
\value{
An object of class 'MaxQuantProteinGroupsResult'.
}
\description{
Load a MaxQuant 'proteinGroups.txt' dataset.
}
